+++ example/lib/main.dart
@@ -1,5 +1,6 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:than_pkg/enums/screen_orientation_types.dart';
 import 'package:than_pkg/than_pkg.dart';
 
 void main() async {
@@ -28,22 +29,15 @@ class _MyAppState extends State<MyApp> {
   }
 
   void init() async {
-    await ThanPkg.android.app.toggleKeepScreenOn(isKeep: true);
+    if (Platform.isAndroid) {
+      await ThanPkg.android.app.toggleKeepScreenOn(isKeep: true);
+    }
   }
 
   void _test() async {
     try {
-      final type = await ThanPkg.android.app.getOrientation();
-      if (type == null) return;
-      if (type == ScreenOrientationTypes.Portrait) {
-        await ThanPkg.android.app
-            .requestOrientation(type: ScreenOrientationTypes.Landscape);
-        await ThanPkg.android.app.showFullScreen();
-      } else {
-        ThanPkg.android.app
-            .requestOrientation(type: ScreenOrientationTypes.Portrait);
-        await ThanPkg.android.app.hideFullScreen();
-      }
+      // ThanPkg.linux.app.launch('/home/thancoder/Videos/【GMV】- BOSS B_TCH.mp4');
+      ThanPkg.platform.launch('/home/thancoder/Videos/【GMV】- BOSS B_TCH.mp4');
     } catch (e) {
       debugPrint(e.toString());
     }
@@ -60,11 +54,7 @@ class _MyAppState extends State<MyApp> {
             : AppBar(
                 title: const Text('test lib'),
               ),
-        body: Center(
-            child: TextField(
-          controller: textEditingController,
-          decoration: InputDecoration(hintText: 'path'),
-        )),
+        body: Placeholder(),
         floatingActionButton: FloatingActionButton(
           onPressed: _test,
           child: Icon(Icons.get_app),