+++ lib/linux_pkg/linux_app.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+class LinuxApp {
+  static final LinuxApp app = LinuxApp._();
+  LinuxApp._();
+  factory LinuxApp() => app;
+
+  Future<void> launch(String source) async {
+    await Process.run('xdg-open', [source]);
+  }
+}