+++ lib/linux_pkg/linux_pkg.dart
@@ -1,3 +1,4 @@
+import 'package:than_pkg/linux_pkg/linux_app.dart';
 import 'package:than_pkg/linux_pkg/linux_thumbnail.dart';
 import 'package:than_pkg/linux_pkg/linux_wifi_util.dart';
 
@@ -8,4 +9,5 @@ class LinuxPkg {
 
   LinuxThumbnail get thumbnail => LinuxThumbnail.thumbnail;
   LinuxWifiUtil get wifi => LinuxWifiUtil.wifi;
+  LinuxApp get app => LinuxApp.app;
 }